Day 9 solution
This one was tricky. The solution to part 2 would actually happen to
work without the check whether all corners of the rectangle are inside
the polygon. This must be a lucky property of the input data AND the
example. It may not hold true for all possible inputs, as inputs vary
by user.

Consider the example:

  ..............
  .......#XXX#..
  .......X...X..
  ..#XXXX#...X..
  ..X........X..
  ..AXXXXXX#.X..
  .........X.X..
  .........BX#..
  ..............

The two "red tiles" marked A and B here span a rectangle that would be
accepted by the algorithm without the corners-in-polygon check. It
happens to be the exact same size of the ACTUAL solution, so it passes
the test. The same property was true for my input data.

But, TECHNICALLY, one must ensure all four corners of the rectangle are
inside the polygon, which would not be the case for the rectangle shown
above.